Separate custom generic bulk views into a separate file
